+++ src/main/java/com/duolingo/open/rtlviewpager/DelegatingPagerAdapter.java
@@ -18,8 +18,8 @@ package com.duolingo.open.rtlviewpager;
 
 import android.database.DataSetObserver;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
-import android.support.v4.view.PagerAdapter;
+import androidx.annotation.NonNull;
+import androidx.viewpager.widget.PagerAdapter;
 import android.view.View;
 import android.view.ViewGroup;